requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pandas==2.2.3
numpy==1.26.4
python-multipart==0.0.9
openpyxl==3.1.5
python-docx==1.1.2
requests==2.32.3
